22_Tools/CNServing/cnserving_sample_code/requirements_client.txt
tensorflow_serving_api==2
pillow
opencv-python
pycocotools
